% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\name{get_tskit}
\alias{get_tskit}
\title{Get the reticulate Python tskit module}
\usage{
get_tskit(obj_name = "tskit")
}
\arguments{
\item{obj_name}{character name of the object holding \code{tskit} reticulate
Python module. If this object exists in the global R environment and is a
reticulate Python object, then it is returned. Otherwise, the function
attempts to install and import tskit before returning it. If \code{NULL},
then the function directly attempts to install and import tskit before
returning it.}
}
\value{
\code{tskit} reticulate Python module.
}
\description{
This function imports the reticulate Python \code{tskit} module
  and if it is not yet installed, then it attempts to install it first.
}
\details{
This function is meant for users running \code{tskit <- get_tskit()}
  or similar, but also by other functions in this package that need the
  \code{tskit} reticulate Python module, yet we don't want to keep importing
  it if it already has been imported.
}
\examples{
tskit <- get_tskit()
is(tskit)
tskit$ALLELES_01
}
