% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tskitr-package.R
\name{ts_dump}
\alias{ts_dump}
\alias{ts_write}
\title{Write a tree sequence to file.}
\usage{
ts_dump(ts, file, options = 0L)

ts_write(ts, file, options = 0L)
}
\arguments{
\item{ts}{an external pointer to a \code{tsk_treeseq_t} object.}

\item{file}{a string specifying the full path of the tree sequence file.}

\item{options}{integer bitwise options (see details).}
}
\description{
Write a tree sequence to file.
}
\details{
This function calls \code{tsk_treeseq_dump()} from the tskit C API
  See \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_dump}
  for more details.
}
\section{Functions}{
\itemize{
\item \code{ts_write()}: Alias for \code{ts_dump()}

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- tskitr::ts_load(ts_file) # slendr also has ts_load()!
ts
dump_file <- tempfile()
ts_dump(ts, dump_file)
}
