% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_load}
\alias{ts_load}
\title{Load tree sequence from a file}
\usage{
ts_load(file)
}
\arguments{
\item{file}{a string specifying the full path of the tree sequence file.}
}
\value{
An external pointer to a \code{tsk_treeseq_t} object.
}
\description{
Load tree sequence from a file
}
\examples{
ts_file <- system.file("examples", "test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts
is(ts)
ts_num_nodes(ts)
}
