% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_num}
\alias{ts_num}
\alias{ts_num_provenances}
\alias{ts_num_populations}
\alias{ts_num_migrations}
\alias{ts_num_individuals}
\alias{ts_num_samples}
\alias{ts_num_nodes}
\alias{ts_num_edges}
\alias{ts_num_trees}
\alias{ts_num_sites}
\alias{ts_num_mutations}
\title{Get the number of records in a tree sequence}
\usage{
ts_num(ts)

ts_num_provenances(ts)

ts_num_populations(ts)

ts_num_migrations(ts)

ts_num_individuals(ts)

ts_num_samples(ts)

ts_num_nodes(ts)

ts_num_edges(ts)

ts_num_trees(ts)

ts_num_sites(ts)

ts_num_mutations(ts)
}
\arguments{
\item{ts}{an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
\code{ts_num} returns a named list with the numbers of each item,
}
\description{
Get the number of records in a tree sequence
}
\details{
These functions return the number of various items in a tree
    sequence, including provenances, populations, migrations, individuals,
    samples, nodes, edges, trees, sites, and mutations.
}
\section{Functions}{
\itemize{
\item \code{ts_num_provenances()}: Get the number of provenances in a tree sequence

\item \code{ts_num_populations()}: Get the number of populations in a tree sequence

\item \code{ts_num_migrations()}: Get the number of migrations in a tree sequence

\item \code{ts_num_individuals()}: Get the number of individuals in a tree sequence

\item \code{ts_num_samples()}: Get the number of samples in a tree sequence

\item \code{ts_num_nodes()}: Get the number of nodes in a tree sequence

\item \code{ts_num_edges()}: Get the number of edges in a tree sequence

\item \code{ts_num_trees()}: Get the number of trees in a tree sequence

\item \code{ts_num_sites()}: Get the number of sites in a tree sequence

\item \code{ts_num_mutations()}: Get the number of mutations in a tree sequence

}}
\examples{
ts_file <- system.file("examples", "test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts_num(ts)
ts_num_provenances(ts)
ts_num_populations(ts)
ts_num_migrations(ts)
ts_num_individuals(ts)
ts_num_samples(ts)
ts_num_nodes(ts)
ts_num_edges(ts)
ts_num_trees(ts)
ts_num_sites(ts)
ts_num_mutations(ts)
}
