% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\name{ts_print}
\alias{ts_print}
\title{Print a summary of a tree sequence and its contents}
\usage{
ts_print(ts)
}
\arguments{
\item{ts}{tree sequence as an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
list with two data.frames; the first contains tree sequence
  properties and their value; the second contains number of rows in
  tables and the length of their metadata.
}
\description{
Print a summary of a tree sequence and its contents
}
\details{
It uses \code{\link{ts_summary}()} and
  \code{\link{ts_metadata_length}()}.
  Note that \code{nbytes} property is not available in
  \code{tskit} C API compared to Python API.
}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts_print(ts)
ts
}
