% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\name{ts_py_to_r}
\alias{ts_py_to_r}
\title{Transfer a tree sequence from reticulate Python to R}
\usage{
ts_py_to_r(ts, cleanup = TRUE)
}
\arguments{
\item{ts}{tree sequence in reticulate Python.}

\item{cleanup}{logical delete the temporary file at the end of the function?}
}
\value{
Tree sequence as an external pointer to a \code{tsk_treeseq_t} object.
}
\description{
This function saves a tree sequence from reticulate Python to disk
  and reads it into R so that we can use \code{tskitr} R/C++ API on it.
}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
# Use the tskit Python API to work with a tree sequence (via reticulate)
tskit <- get_tskit_py()
ts_py <- tskit$load(ts_file)
is(ts_py)
ts_py$num_samples # 160
ts2_py <- ts_py$simplify(samples = c(0L, 1L, 2L, 3L))
ts2_py$num_samples # 4
# Now transfer the tree sequence to R and use tskitr R/C++ API on it
ts2_r <- ts_py_to_r(ts2_py)
is(ts2_r)
ts_num_samples(ts2_r) # 4
}
\seealso{
\code{\link{ts_r_to_py}()} and \code{\link{ts_dump}()}.
}
