% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_summary}
\alias{ts_summary}
\alias{ts_num_provenances}
\alias{ts_num_populations}
\alias{ts_num_migrations}
\alias{ts_num_individuals}
\alias{ts_num_samples}
\alias{ts_num_nodes}
\alias{ts_num_edges}
\alias{ts_num_trees}
\alias{ts_num_sites}
\alias{ts_num_mutations}
\alias{ts_sequence_length}
\alias{ts_time_units}
\title{Get the summary of properties and number of records in a tree
  sequence}
\usage{
ts_summary(ts)

ts_num_provenances(ts)

ts_num_populations(ts)

ts_num_migrations(ts)

ts_num_individuals(ts)

ts_num_samples(ts)

ts_num_nodes(ts)

ts_num_edges(ts)

ts_num_trees(ts)

ts_num_sites(ts)

ts_num_mutations(ts)

ts_sequence_length(ts)

ts_time_units(ts)
}
\arguments{
\item{ts}{tree sequence as an external pointer to a \code{tsk_treeseq_t}
object.}
}
\value{
\code{ts_summary} returns a named list with the number/value for
  all items, while other functions \code{ts_num_x} etc. return the
  number/value of each item.
}
\description{
Get the summary of properties and number of records in a tree
  sequence
}
\details{
These functions return the summary of properties and number of
  records in a tree sequence, by calling
  \code{tsk_treeseq_get_num_provenances()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_provenances},
  \code{tsk_treeseq_get_num_populations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_populations},
  \code{tsk_treeseq_get_num_migrations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_migrations},
  \code{tsk_treeseq_get_num_individuals()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_individuals},
  \code{tsk_treeseq_get_num_samples()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_samples},
  \code{tsk_treeseq_get_num_nodes()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_nodes},
  \code{tsk_treeseq_get_num_edges()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_edges},
  \code{tsk_treeseq_get_num_trees()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_trees},
  \code{tsk_treeseq_get_num_sites()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_sites},
  \code{tsk_treeseq_get_num_mutations()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_num_mutations},
  \code{tsk_treeseq_get_sequence_length()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_sequence_length},
  and
  \code{tsk_treeseq_get_time_units()}
\url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_time_units}.
  from the tskit C API. See the linked documentation for more details.
}
\section{Functions}{
\itemize{
\item \code{ts_num_provenances()}: Get the number of provenances in a tree sequence

\item \code{ts_num_populations()}: Get the number of populations in a tree sequence

\item \code{ts_num_migrations()}: Get the number of migrations in a tree sequence

\item \code{ts_num_individuals()}: Get the number of individuals in a tree sequence

\item \code{ts_num_samples()}: Get the number of samples (of nodes) in a tree
sequence

\item \code{ts_num_nodes()}: Get the number of nodes in a tree sequence

\item \code{ts_num_edges()}: Get the number of edges in a tree sequence

\item \code{ts_num_trees()}: Get the number of trees in a tree sequence

\item \code{ts_num_sites()}: Get the number of sites in a tree sequence

\item \code{ts_num_mutations()}: Get the number of mutations in a tree sequence

\item \code{ts_sequence_length()}: Get the sequence length

\item \code{ts_time_units()}: Get the time units string

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- tskitr::ts_load(ts_file) # slendr also has ts_load()!
ts_summary(ts)
ts_num_provenances(ts)
ts_num_populations(ts)
ts_num_migrations(ts)
ts_num_individuals(ts)
ts_num_samples(ts)
ts_num_nodes(ts)
ts_num_edges(ts)
ts_num_trees(ts)
ts_num_sites(ts)
ts_num_mutations(ts)
ts_sequence_length(ts)
ts_time_units(ts)
}
