% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\docType{package}
\name{tskitr-package}
\alias{tskitr}
\alias{tskitr-package}
\title{tskitr: R access to the tskit C API}
\description{
Tskit enables performant storage, manipulation and analysis of ancestral
recombination graphs using succinct tree sequence encoding; see https://tskit.dev.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` package to seamlessly load and analyse tree sequences;
see https://tskit.dev/tutorials/tskitr.html.
`tskitr` provides R access to the tskit C API for use cases where the
`reticulate` approach is not suitable. For example, where high-performance
and low-level work with tree sequences is required. R access to the parts of
C API is added as the need arises.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/tskitr}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
