% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tskitr-package.R
\name{ts_load}
\alias{ts_load}
\alias{ts_read}
\title{Load a tree sequence from a file.}
\usage{
ts_load(file, options = 0L)

ts_read(file, options = 0L)
}
\arguments{
\item{file}{a string specifying the full path of the tree sequence file.}

\item{options}{integer bitwise options (see details).}
}
\value{
Tree sequence as an external pointer to a \code{tsk_treeseq_t}
  object.
}
\description{
Load a tree sequence from a file.
}
\details{
This function calls \code{tsk_treeseq_load()} from the tskit C API.
  \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_load}
  provides more details.
}
\section{Functions}{
\itemize{
\item \code{ts_read()}: Alias for \code{ts_load()}

}}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts
is(ts)
ts_num_nodes(ts)
}
