% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_metadata_length}
\alias{ts_metadata_length}
\title{Get the length of metadata in a tree sequence and its tables}
\usage{
ts_metadata_length(ts)
}
\arguments{
\item{ts}{tree sequence as an external pointer to a \code{tsk_treeseq_t} object.}
}
\value{
A named list with the length of metadata.
}
\description{
Get the length of metadata in a tree sequence and its tables

from the tskit C API. See the linked documentation for more details.
}
\details{
This function returns the length of metadata stored on the tree
  sequence and in each table by calling
  \code{tsk_treeseq_get_metadata_length()} \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_get_metadata_length} on tree sequence and
}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- tskitr::ts_load(ts_file) # slendr also has ts_load()!
ts_metadata_length(ts)
}
