% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\name{ts_py_to_r}
\alias{ts_py_to_r}
\title{Transfer a tree sequence from reticulate Python to R}
\usage{
ts_py_to_r(ts, cleanup = TRUE)
}
\arguments{
\item{ts}{tree sequence in reticulate Python.}

\item{cleanup}{logical delete the temporary file at the end of the function?}
}
\value{
Tree sequence as an external pointer to a \code{tsk_treeseq_t} object.
}
\description{
This function saves a tree sequence from reticulate Python to disk
  and reads it into R so that we can use \code{tskitr} R/C++ API on it.
}
\examples{
ts_file <- system.file("examples/test.trees", package = "tskitr")
# Use the tskit Python API to work with a tree sequence (via reticulate)
tskit <- get_tskit()
py_ts <- tskit$load(ts_file)
is(py_ts)
py_ts$num_samples # 160
py_ts2 <- py_ts$simplify(samples = c(0L, 1L, 2L, 3L))
py_ts2$num_samples # 4
# Now transfer the tree sequence to R and use tskitr R/C++ API on it
r_ts2 <- ts_py_to_r(py_ts2)
is(r_ts2)
ts_num_samples(r_ts2) # 4
}
\seealso{
\code{\link{ts_r_to_py}()} and \code{\link{ts_dump}()}.
}
