% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\docType{package}
\name{tskitr-package}
\alias{tskitr}
\alias{tskitr-package}
\title{tskitr: R access to the tskit C API}
\description{
Tskit (`tskit`)  enables performant storage, manipulation, and analysis of
ancestral recombination graphs using succinct tree sequence encoding; see
https://tskit.dev.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` R package to seamlessly load and analyse a tree sequence;
see https://tskit.dev/tutorials/tskitr.html.
`tskitr` provides R access to the tskit C API for use cases where the
`reticulate` approach is not optimal. For example, for high-performance
and low-level work with tree sequences. Currently, `tskitr` provides a very
limited number of R functions due to the availability of extensive Python API
and the `reticulate` approach.
}
\examples{
\dontshow{# Providing the examples here so we test them via R CMD check}
# Here are examples showcasing what you can do with tskitr

# 1) Load a tree sequence into an R session and summarise its contents
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts_num_individuals(ts)

# 2) Call tskit C API in C++ code in an R session
codeString <- '
  #include <tskit.h>
  int ts_num_individuals(SEXP ts) {
    int n;
    Rcpp::XPtr<tsk_treeseq_t> xptr(ts);
    n = (int) tsk_treeseq_get_num_individuals(xptr);
    return n;
  }'
ts_num_individuals2 <- Rcpp::cppFunction(code=codeString, depends="tskitr", plugins="tskitr")
ts_file <- system.file("examples/test.trees", package="tskitr")
ts <- tskitr::ts_load(ts_file) # slendr also has ts_load()!
ts_num_individuals2(ts)
ts_num_individuals(ts) # tskitr implementation of ts_num_individuals2()

# 3) Call `tskit` C API in C++ code in another R package
state_and_aims_file <- system.file("STATE_and_AIMS.md", package = "tskitr")
browseURL(state_and_aims_file)

# 4) Call `tskit` C API in R code in an R session or another R package
# TODO: Write STATE_and_AIMS.md file on what we want to achieve #12
#       https://github.com/HighlanderLab/tskitr/issues/12
state_and_aims_file <- system.file("STATE_and_AIMS.md", package = "tskitr")
browseURL(state_and_aims_file)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/tskitr}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
