% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\docType{package}
\name{tskitr-package}
\alias{tskitr}
\alias{tskitr-package}
\title{tskitr: R access to the tskit C API}
\description{
Tskit enables performant storage, manipulation and analysis of ancestral
recombination graphs using succinct tree sequence encoding; see https://tskit.dev.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` package to seamlessly load and analyse tree sequences;
see https://tskit.dev/tutorials/tskitr.html.
`tskitr` provides R access to the tskit C API for use cases where the
`reticulate` approach is not suitable. For example, where high-performance
and low-level work with tree sequences is required. R access to the parts of
C API is added as the need arises.
}
\examples{
# Here are two examples showcasing what you can do with tskitr

# 1) Load a tree sequence into an R session and summarise its contents
ts_file <- system.file("examples", "test.trees", package = "tskitr")
ts <- ts_load(ts_file)
ts_num_individuals(ts)

# 2) Call tskit C API in C++ code within an R session or an R package
library(Rcpp)
codeString <- '
    #include <tskit.h>
    int ts_num_individuals(SEXP ts) {
        int n;
        Rcpp::XPtr<tsk_treeseq_t> xptr(ts);
        n = (int) tsk_treeseq_get_num_individuals(xptr);
        return n;
    }'
get_num_individuals <- cppFunction(code=codeString, depends="tskitr", plugins="tskitr")
ts_file <- system.file("examples", "test.trees", package="tskitr")
ts <- tskitr::ts_load(ts_file)
get_num_individuals(ts)
ts_num_individuals(ts) # tskitr implementation of get_num_individuals()

# For use in an R package, see the vignette TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/tskitr}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
