% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskitr-package.R
\docType{package}
\name{tskitr-package}
\alias{tskitr}
\alias{tskitr-package}
\title{tskitr: R access to the tskit C API}
\description{
Tskit enables performant storage, manipulation, and analysis of
ancestral recombination graphs using succinct tree sequence encoding.
See https://tskit.dev for project news, documentation, and tutorials.
Tskit provides Python, C, and Rust APIs. The Python API can be called from R
via the `reticulate` R package to seamlessly load and analyse a tree sequence
as described at https://tskit.dev/tutorials/tskitr.html.
`tskitr` provides R access to the tskit C API for use cases where the
`reticulate` approach is not optimal. For example, for high-performance
and low-level work with tree sequences. Currently, `tskitr` provides a very
limited number of R functions due to the availability of extensive Python API
and the `reticulate` approach.
}
\examples{
\dontshow{# Providing the examples here so we test them via R CMD check}
# Here are examples showcasing what you can do with tskitr

# 1) Load a tree sequence into an R session and summarise it
# Load a tree sequence
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- ts_load(ts_file)

# Print summary of the tree sequence
ts_num_individuals(ts)
ts_print(ts)

# 2) Pass tree sequence between R and reticulate or standard Python

# Tree sequence in R
ts_file <- system.file("examples/test.trees", package = "tskitr")
ts <- ts_load(ts_file)

# If you have a tree sequence in R and you want to use tskit Python API,
# you can write it to disk and load it into a reticulate Python session
ts_py <- ts_r_to_py(ts)
# ... continue in reticulate Python ...
ts_py$num_individuals # 160
ts2_py = ts_py$simplify(samples = c(0L, 1L, 2L, 3L))
ts2_py$num_individuals # 2
# ... and to bring it back to R ...
ts2 <- ts_py_to_r(ts2_py)
ts_num_individuals(ts2) # 2

# If you prefer a standard (non-reticulate) Python, use this
ts_file <- tempfile()
print(ts_file)
ts_dump(ts, file = ts_file)
# ... continue in a standard Python session ...
# import tskit
# ts = tskit.load("insert_ts_file_path_here")
# ts.num_individuals # 80
# ts2 = ts.simplify(samples = [0, 1, 2, 3])
# ts2.num_individuals # 2
# ts2.dump("insert_ts_file_path_here")
# ... and to bring it back to R ...
ts2 <- ts_load(ts_file)
ts_num_individuals(ts2) # 2
file.remove(ts_file)

# 3) Call tskit C API in C++ code in an R session
library(Rcpp)
# Write and compile a C++ function
codeString <- '
  #include <tskit.h>
  int ts_num_individuals(SEXP ts) {
    int n;
    Rcpp::XPtr<tsk_treeseq_t> xptr(ts);
    n = (int) tsk_treeseq_get_num_individuals(xptr);
    return n;
  }'
ts_num_individuals2 <- Rcpp::cppFunction(code=codeString, depends="tskitr", plugins="tskitr")
# We must specify both the `depends` and `plugins` arguments!

# Load a tree sequence
ts_file <- system.file("examples/test.trees", package="tskitr")
ts <- ts_load(ts_file)

# Apply the compiled function
ts_num_individuals2(ts)

# An identical tskitr implementation
ts_num_individuals(ts)

# 4) Call `tskit` C API in C++ code in another R package
# TODO: add vignette issue link here
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/highlanderlab/tskitr}
}

}
\author{
\strong{Maintainer}: Gregor Gorjanc \email{gregor.gorjanc@gmail.com} (\href{https://orcid.org/0000-0001-8008-2787}{ORCID})

}
\keyword{internal}
